fix: apply pr#7 reviews
