+++ src/pages/create/index.tsx
@@ -4,7 +4,7 @@ import { css } from '@emotion/react';
 import { mediaQuery, pxToVw } from 'styles/media';
 import TextInput from 'components/text-input';
 import Button from 'components/button';
-import DatePicker from 'components/date-picker';
+import DateSelector from 'components/date-selector';
 
 const button = css`
   margin-top: ${pxToVw(40)};
@@ -15,10 +15,12 @@ const button = css`
 `;
 
 export default function Create() {
+  const defaultDate = new Date();
+
   return (
     <div css={basicWrap}>
       <TextInput placeholder="여행 이름 입력" note="이름은 최소 0자, 최대 00자까지 입력 가능해요" />
-      <DatePicker />
+      <DateSelector defaultDate={defaultDate} />
       <Button label="다음" customStyle={button} />
     </div>
   );
